Refactor zet edit and search functions, upgrade dependencies
Refactored the zet edit and search functions to reduce code redundancy by using 'z.Path' instead of local variables. Removed string return in search function and the usage of the variable 'zet' in the render and edit methods. Also, updated libraries to a newer version and bumped the application version to 'v0.5.0'.

Signed-off-by: Daniel Michaels <dan@danielms.site>